chore(richtext-lexical): make sure no client-side code is imported by default from Features
